admin

+++ src/main/java/Pack01/AdminCon.java
@@ -1,5 +1,9 @@
 package Pack01;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -9,19 +13,25 @@ import org.springframework.web.bind.annotation.RequestParam;
 @Controller
 public class AdminCon {
 	@RequestMapping(value = "/admin", method = RequestMethod.GET)
-	String adminPage() {
+	String adminPage(HttpServletRequest request, HttpServletResponse response) {
+		HttpSession session = request.getSession();
+		if(!(session.getAttribute("cn").equals("admin"))) return "logout";
 		return "admin";
 	}
 	
 	@RequestMapping(value = "/admin/listexam", method = RequestMethod.GET)
-	String listExam(Model model) {
+	String listExam(Model model, HttpServletRequest request, HttpServletResponse response) {
+		HttpSession session = request.getSession();
+		if(!(session.getAttribute("cn").equals("admin"))) return "logout";
 		System.out.println("문제보기");
 		model.addAttribute("questionList", AdminDAO.examList());
 		return "ListExam";
 	}
 	
 	@RequestMapping(value = "/admin/addexam", method = RequestMethod.GET)
-	String addExamView() {
+	String addExamView(HttpServletRequest request, HttpServletResponse response) {
+		HttpSession session = request.getSession();
+		if(!(session.getAttribute("cn").equals("admin"))) return "logout";
 		System.out.println("문제추가페이지");
 		return "AddExam";
 	}
@@ -33,8 +43,11 @@ public class AdminCon {
 			@RequestParam(value="ans_2") String ans_2,
 			@RequestParam(value="ans_3") String ans_3,
 			@RequestParam(value="ans_4") String ans_4,
-			@RequestParam(value="answer") String answer
+			@RequestParam(value="answer") String answer,
+			HttpServletRequest request, HttpServletResponse response
 			) {
+		HttpSession session = request.getSession();
+		if(!(session.getAttribute("cn").equals("admin"))) return "logout";
 		Boolean daoCheck = AdminDAO.InsertProblem(question, ans_1, ans_2, ans_3, ans_4, answer);
 		if (!daoCheck) {
 			return "admin";
@@ -43,15 +56,20 @@ public class AdminCon {
 	}
 	
 	@RequestMapping(value = "/admin/result", method = RequestMethod.GET)
-	String readResult(Model model) {
+	String readResult(Model model, HttpServletRequest request, HttpServletResponse response) {
+		HttpSession session = request.getSession();
+		if(!(session.getAttribute("cn").equals("admin"))) return "logout";
 		System.out.println("시험 본 결과");
 		model.addAttribute("resultList", AdminDAO.ResultList());
 		return "ResultList";
 	}
 	
 	@RequestMapping(value = "/admin/examupdate", method = RequestMethod.GET)
-	String updateExamView(Model model, @RequestParam(value="id") String updateId) {
+	String updateExamView(Model model, @RequestParam(value="id") String updateId,
+			HttpServletRequest request, HttpServletResponse response) {
 		System.out.println("시험 수정 view");
+		HttpSession session = request.getSession();
+		if(!(session.getAttribute("cn").equals("admin"))) return "logout";
 		model.addAttribute("exam", AdminDAO.ExamGet(updateId));
 		return "UpdateExam";
 	}
@@ -64,9 +82,12 @@ public class AdminCon {
 			@RequestParam(value="ans_2") String ans_2,
 			@RequestParam(value="ans_3") String ans_3,
 			@RequestParam(value="ans_4") String ans_4,
-			@RequestParam(value="answer") String answer
+			@RequestParam(value="answer") String answer,
+			HttpServletRequest request, HttpServletResponse response
 			) {
 		System.out.println("시험 수정");
+		HttpSession session = request.getSession();
+		if(!(session.getAttribute("cn").equals("admin"))) return "logout";
 		Boolean daoCheck = AdminDAO.UpdateExam(updateId, question, ans_1, ans_2, ans_3, ans_4, answer);
 		if(daoCheck) {
 			return "redirect:/admin/listexam";
@@ -75,7 +96,10 @@ public class AdminCon {
 	}
 	
 	@RequestMapping(value = "/admin/examdelete", method = RequestMethod.GET)
-	String deleteExam(@RequestParam(value="id") String deleteId) {
+	String deleteExam(@RequestParam(value="id") String deleteId,
+			HttpServletRequest request, HttpServletResponse response) {
+		HttpSession session = request.getSession();
+		if(!(session.getAttribute("cn").equals("admin"))) return "logout";
 		Boolean daoCheck = AdminDAO.DeleteExam(deleteId);
 		if(daoCheck) {
 			return "redirect:/admin/listexam";
